feat!: remove legacy scans
