+++ bec_server/bec_server/scan_server/scan_assembler.py
@@ -2,26 +2,44 @@ from __future__ import annotations
 
 import inspect
 from functools import partial
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any
 
 from bec_lib import messages
-from bec_lib.device import DeviceBase
 from bec_lib.logger import bec_logger
 from bec_lib.scan_input_validator import ScanInputValidator
 from bec_lib.signature_serializer import serialize_dtype
 
-from .scans.legacy_scans import RequestBase, ScanArgType, ScanBase, unpack_scan_args
 from .scans.scan_argument_modifier import (
     apply_scan_argument_defaults,
     get_scan_modifier,
     scan_signature_with_modifiers,
 )
-from .scans.scan_base import ScanBase as ScanBaseV4
+from .scans.scan_base import ScanBase
 
 logger = bec_logger.logger
 
 if TYPE_CHECKING:
     from .scan_server import ScanServer
+
+
+def unpack_scan_args(scan_args: dict[str, Any]) -> list:
+    """unpack_scan_args unpacks the scan arguments and returns them as a tuple.
+
+    Args:
+        scan_args (dict[str, Any]): scan arguments
+
+    Returns:
+        list: list of arguments
+    """
+    args = []
+    if not scan_args:
+        return args
+    if not isinstance(scan_args, dict):
+        return scan_args
+    for cmd_name, cmd_args in scan_args.items():
+        args.append(cmd_name)
+        args.extend(cmd_args)
+    return args
 
 
 class ScanAssembler:
@@ -36,19 +54,6 @@ class ScanAssembler:
         self.scan_manager = self.parent.scan_manager
         self.input_validator = ScanInputValidator(device_manager=self.device_manager)
 
-    def is_scan_message(self, msg: messages.ScanQueueMessage) -> bool:
-        """Check if the scan queue message would construct a new scan.
-
-        Args:
-            msg (messages.ScanQueueMessage): message to be checked
-
-        Returns:
-            bool: True if the message is a scan message, False otherwise
-        """
-        scan = msg.content.get("scan_type")
-        scan_cls = self.scan_manager.scan_dict[scan]
-        return issubclass(scan_cls, ScanBase)
-
     def is_direct_scan_message(self, msg: messages.ScanQueueMessage) -> bool:
         """Check if the scan queue message would construct a new direct scan.
 
@@ -59,50 +64,11 @@ class ScanAssembler:
         """
         scan = msg.content.get("scan_type")
         scan_cls = self.scan_manager.scan_dict[scan]
-        return issubclass(scan_cls, ScanBaseV4)
+        return issubclass(scan_cls, ScanBase)
 
-    def assemble_device_instructions(
-        self, msg: messages.ScanQueueMessage, scan_id: str
-    ) -> RequestBase:
+    def assemble_direct_scan(self, msg: messages.ScanQueueMessage, scan_id: str | None) -> ScanBase:
         """Assemble the device instructions for a given ScanQueueMessage.
-        This will be achieved by calling the specified class (must be a derived class of RequestBase)
-
-        Args:
-            msg (messages.ScanQueueMessage): scan queue message for which the instruction should be assembled
-            scan_id (str): scan id of the scan
-
-        Raises:
-            ScanAbortion: Raised if the scan initialization fails.
-
-        Returns:
-            RequestBase: Scan instance of the initialized scan class
-        """
-        scan = msg.content.get("scan_type")
-        scan_cls = self.scan_manager.scan_dict[scan]
-
-        logger.info(f"Preparing instructions of request of type {scan} / {scan_cls.__name__}")
-        args = unpack_scan_args(msg.content.get("parameter", {}).get("args", []))
-        kwargs = msg.content.get("parameter", {}).get("kwargs", {})
-
-        request_inputs = self._assemble_request_inputs(scan_cls, args, kwargs)
-
-        scan_instance = scan_cls(
-            *args,
-            device_manager=self.device_manager,
-            parameter=msg.content.get("parameter"),
-            metadata=msg.metadata,
-            instruction_handler=self.parent.queue_manager.instruction_handler,
-            scan_id=scan_id,
-            request_inputs=request_inputs,
-            **kwargs,
-        )
-        return scan_instance
-
-    def assemble_direct_scan(
-        self, msg: messages.ScanQueueMessage, scan_id: str | None
-    ) -> ScanBaseV4:
-        """Assemble the device instructions for a given ScanQueueMessage.
-        This will be achieved by calling the specified class (must be a derived class of ScanBaseV4)
+        This will be achieved by calling the specified class (must be a derived class of ScanBase)
 
         Args:
             msg (messages.ScanQueueMessage): scan queue message for which the instruction should be assembled
@@ -112,7 +78,7 @@ class ScanAssembler:
             ScanAbortion: Raised if the scan initialization fails.
 
         Returns:
-            ScanBaseV4: Scan instance of the initialized scan class
+            ScanBase: Scan instance of the initialized scan class
         """
         scan = msg.content.get("scan_type")
         scan_cls = self.scan_manager.scan_dict[scan]
@@ -217,21 +183,5 @@ class ScanAssembler:
     def _serialize_arg_input(self, arg_input: dict) -> dict[str, str | dict | list]:
         converted = {}
         for key, value in arg_input.items():
-            if value == ScanArgType.DEVICE:
-                value = DeviceBase
-            elif value == ScanArgType.FLOAT:
-                value = float
-            elif value == ScanArgType.INT:
-                value = int
-            elif value == ScanArgType.BOOL:
-                value = bool
-            elif value == ScanArgType.STR:
-                value = str
-            elif value == ScanArgType.LIST:
-                value = list
-            elif value == ScanArgType.DICT:
-                value = dict
-            elif inspect.isclass(value) and issubclass(value, DeviceBase):
-                value = DeviceBase
             converted[key] = serialize_dtype(value)
         return converted
