+++ bec_server/bec_server/scan_server/scan_queue.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import collections
 import functools
 import threading
-import time
 import traceback
 import uuid
 from enum import Enum
@@ -16,9 +15,7 @@ from bec_lib import messages
 from bec_lib.alarm_handler import Alarms
 from bec_lib.endpoints import MessageEndpoints
 from bec_lib.logger import bec_logger
-from bec_server.scan_server.types import ReadoutPriorities
 
-from .errors import DeviceInstructionError, LimitError, ScanAbortion
 from .instruction_handler import InstructionHandler
 from .scan_assembler import ScanAssembler
 
@@ -180,7 +177,6 @@ class QueueManager:
     def _scan_queue_callback(self, msg) -> None:
         scan_msg = msg.value
         logger.info(f"Receiving scan: {scan_msg.content}")
-        # instructions = self.scan_assembler.assemble_device_instructions(scan_msg)
         queue = scan_msg.content.get("queue", "primary")
         self.add_to_queue(queue, scan_msg)
 
@@ -251,7 +247,7 @@ class QueueManager:
 
     def _get_queue_item_by_scan_id(
         self, msg: messages.ScanQueueOrderMessage
-    ) -> InstructionQueueItem | DirectInstructionQueueItem | None:
+    ) -> DirectInstructionQueueItem | None:
         """
         Get the queue item by scan_id.
 
@@ -425,9 +421,7 @@ class QueueManager:
                 devices=self._get_owned_devices_for_instruction_queue(instruction_queue),
             )
 
-    def _cancel_queue_item(
-        self, target_queue_item: InstructionQueueItem | DirectInstructionQueueItem, queue: str
-    ) -> None:
+    def _cancel_queue_item(self, target_queue_item: DirectInstructionQueueItem, queue: str) -> None:
         """
         Mark a pending queue item as cancelled before removing it from the queue.
         This is to allow clients to recognize that the scan was cancelled and did not just
@@ -454,10 +448,7 @@ class QueueManager:
         exit_info = ("halted", "user" if user_call else "alarm")
         instruction_queue = self.queues[queue].active_instruction_queue
         if instruction_queue:
-            if isinstance(instruction_queue, DirectInstructionQueueItem):
-                instruction_queue.run_on_exception_hook = False
-            else:
-                instruction_queue.return_to_start = False
+            instruction_queue.run_on_exception_hook = False
         self.set_abort(scan_id=scan_id, request_id=request_id, queue=queue, exit_info=exit_info)
 
     @requires_queue
@@ -609,7 +600,7 @@ class QueueManager:
 
     def _get_queue_item_by_request_id(
         self, queue: str, request_id: str
-    ) -> InstructionQueueItem | DirectInstructionQueueItem | None:
+    ) -> DirectInstructionQueueItem | None:
         for instruction_queue in self.queues[queue].queue:
             request_blocks = instruction_queue.describe().request_blocks
             if any(request_block.RID == request_id for request_block in request_blocks):
@@ -629,49 +620,18 @@ class QueueManager:
         return instr_queue.active_request_block.scan_id
 
     def _get_owned_devices_for_instruction_queue(
-        self, instruction_queue: InstructionQueueItem | DirectInstructionQueueItem
+        self, instruction_queue: DirectInstructionQueueItem
     ) -> list[str] | None:
         registry = getattr(self.parent, "device_lock_registry", None)
         if registry is None:
-            return None if isinstance(instruction_queue, InstructionQueueItem) else []
-        if isinstance(instruction_queue, DirectInstructionQueueItem):
-            if instruction_queue.active_scan is None:
-                return []
-            request_id = instruction_queue.active_scan.scan_info.metadata.get("RID")
-        else:
-            if instruction_queue.active_request_block is None:
-                return None
-            request_id = instruction_queue.active_request_block.RID
+            return []
+        if instruction_queue.active_scan is None:
+            return []
+        request_id = instruction_queue.active_scan.scan_info.metadata.get("RID")
         if request_id is None:
-            return None if isinstance(instruction_queue, InstructionQueueItem) else []
+            return []
         devices = registry.get_owned_devices(request_id)
-        if isinstance(instruction_queue, InstructionQueueItem) and not devices:
-            return None
         return devices
-
-    def _wait_for_queue_to_appear_in_history(
-        self, scan_id, queue, timeout=60
-    ) -> InstructionQueueItem:
-        timeout_time = timeout
-        elapsed_time = 0
-        while True:
-            if elapsed_time > timeout_time:
-                raise TimeoutError(
-                    f"Scan {scan_id} did not appear in history within {timeout_time}s"
-                )
-            elapsed_time += 0.1
-            history = self.queues[queue].history_queue
-            if len(history) == 0:
-                time.sleep(0.1)
-                continue
-            if scan_id not in history[-1].scan_id:
-                time.sleep(0.1)
-                continue
-
-            if len(self.queues[queue].queue) > 0 and scan_id in self.queues[queue].queue[0].scan_id:
-                time.sleep(0.1)
-                continue
-            return history[-1]
 
     def send_queue_status(self) -> None:
         """send the current queue to redis"""
@@ -761,15 +721,13 @@ class ScanQueue:
         self,
         queue_manager: QueueManager,
         queue_name="primary",
-        instruction_queue_item_cls: (
-            type[InstructionQueueItem] | type[DirectInstructionQueueItem] | None
-        ) = None,
+        instruction_queue_item_cls: type[DirectInstructionQueueItem] | None = None,
     ) -> None:
-        self.queue: Deque[InstructionQueueItem | DirectInstructionQueueItem] = collections.deque()
+        self.queue: Deque[DirectInstructionQueueItem] = collections.deque()
         self._deferred_inserts: Deque[tuple[messages.ScanQueueMessage, int]] = collections.deque()
         self.queue_name = queue_name
-        self.history_queue: collections.deque[InstructionQueueItem | DirectInstructionQueueItem] = (
-            collections.deque(maxlen=self.MAX_HISTORY)
+        self.history_queue: collections.deque[DirectInstructionQueueItem] = collections.deque(
+            maxlen=self.MAX_HISTORY
         )
         self.active_instruction_queue = None
         self.queue_manager = queue_manager
@@ -1025,12 +983,7 @@ class ScanQueue:
                 queue_exists = True
         if not queue_exists:
             # create new queue element (InstructionQueueItem)
-            assembler = self.queue_manager.parent.scan_assembler
-            if assembler.is_direct_scan_message(msg):
-                iq_class = DirectInstructionQueueItem
-            else:
-                iq_class = InstructionQueueItem
-            iq_class = self._instruction_queue_item_cls_override or iq_class
+            iq_class = self._instruction_queue_item_cls_override or DirectInstructionQueueItem
 
             instruction_queue = iq_class(
                 parent=self,
@@ -1087,15 +1040,6 @@ class ScanQueue:
         if self.active_instruction_queue is not None:
             self.active_instruction_queue.abort()
 
-    def get_scan(self, scan_id: str) -> InstructionQueueItem | None:
-        """get the instruction queue item based on its scan_id"""
-        queue_found = None
-        for queue in self.history_queue + self.queue:
-            if queue.scan_id == scan_id:
-                queue_found = queue
-                return queue_found
-        return queue_found
-
 
 class AutoResetCM:
     """Context manager to automatically reset the queue status"""
@@ -1110,451 +1054,6 @@ class AutoResetCM:
     def __exit__(self, exc_type, exc_value, exc_traceback):
         self.queue.auto_reset_enabled = True
         return False
-
-
-class RequestBlock:
-    def __init__(
-        self, msg: messages.ScanQueueMessage, assembler: ScanAssembler, parent: RequestBlockQueue
-    ) -> None:
-        self.instructions = None
-        self.readout_priority: ReadoutPriorities = {}
-        self.msg = msg
-        self.RID = msg.metadata["RID"]
-        self.scan_assembler = assembler
-        self.scan_id = None
-        self.is_scan = False
-        self._scan_number = None
-        self.parent = parent
-        self._assemble()
-        self.scan_report_instructions = []
-
-    def _assemble(self):
-        self.is_scan = self.scan_assembler.is_scan_message(self.msg)
-        if self.is_scan or self.scan_def_id is not None:
-            self.scan_id = str(uuid.uuid4())
-        self.scan = self.scan_assembler.assemble_device_instructions(self.msg, self.scan_id)
-        self.instructions = self.scan.run()
-        self.readout_priority = self.scan.readout_priority
-
-    @property
-    def scan_def_id(self):
-        return self.msg.metadata.get("scan_def_id")
-
-    @property
-    def metadata(self):
-        return self.msg.metadata
-
-    @property
-    def scan_number(self):
-        """get the predicted scan number"""
-        if not self.is_scan:
-            return None
-        if self._scan_number is not None:
-            return self._scan_number
-        return self._scan_server_scan_number + self.scan_ids_head()
-
-    @property
-    def _scan_server_scan_number(self):
-        return self.parent.scan_queue.queue_manager.parent.scan_number
-
-    def assign_scan_number(self):
-        """assign and fix the current scan number prediction"""
-        if self.parent is None:
-            return
-
-        if not self.is_scan and self.msg.scan_type not in [
-            "open_scan_def",
-            "_open_interactive_scan",
-        ]:
-            return
-        if self.is_scan and self.scan_def_id is not None:
-            return
-        with self.parent.scan_queue.queue_manager._lock:
-            self.parent.increase_scan_number()
-            self._scan_number = self._scan_server_scan_number
-            if hasattr(self.scan, "scan_number"):
-                self.scan.scan_number = self._scan_number
-        return
-
-    def scan_ids_head(self) -> int:
-        """calculate the scan_id offset in the queue for the current request block"""
-        offset = 1
-        for queue in self.parent.scan_queue.queue:
-            if queue.status in [InstructionQueueStatus.COMPLETED, InstructionQueueStatus.RUNNING]:
-                continue
-            if queue.queue_id != self.parent.instruction_queue.queue_id:
-                offset += len([scan_id for scan_id in queue.scan_id if scan_id])
-            else:
-                for scan_id in queue.scan_id:
-                    if scan_id == self.scan_id:
-                        return offset
-                    if scan_id:
-                        offset += 1
-                return offset
-        return offset
-
-    def describe(self) -> messages.RequestBlock:
-        """prepare a dictionary that summarizes the request block"""
-        return messages.RequestBlock(
-            msg=self.msg,
-            RID=self.RID,
-            readout_priority=self.readout_priority,
-            is_scan=self.is_scan,
-            scan_number=self.scan_number,
-            scan_id=self.scan_id,
-            report_instructions=self.scan_report_instructions,
-            owned_device_locks=[],
-            pending_device_locks=[],
-        )
-
-
-class RequestBlockQueue:
-    def __init__(self, instruction_queue, assembler) -> None:
-        self.request_blocks_queue = collections.deque()
-        self.request_blocks: list[RequestBlock] = []
-        self.instruction_queue = instruction_queue
-        self.scan_queue = instruction_queue.parent
-        self.assembler = assembler
-        self.active_rb = None
-        self.scan_def_ids = {}
-
-    @property
-    def scan_id(self) -> list[str]:
-        """get the scan_ids for all request blocks"""
-        return [rb.scan_id for rb in self.request_blocks]
-
-    @property
-    def is_scan(self) -> list[bool]:
-        """check if the request blocks describe scans"""
-        return [rb.is_scan for rb in self.request_blocks]
-
-    @property
-    def scan_number(self) -> list[int]:
-        """get the list of scan numbers for all request blocks"""
-        return [rb.scan_number for rb in self.request_blocks]
-
-    def append(self, msg: messages.ScanQueueMessage) -> None:
-        """append a new scan queue message"""
-        request_block = RequestBlock(msg, self.assembler, parent=self)
-        self._update_scan_def_id(request_block)
-        self.append_request_block(request_block)
-
-    def _update_scan_def_id(self, request_block: RequestBlock):
-        if "scan_def_id" not in request_block.msg.metadata:
-            return
-        scan_def_id = request_block.msg.metadata["scan_def_id"]
-        if scan_def_id in self.scan_def_ids:
-            request_block.scan_id = self.scan_def_ids[scan_def_id]["scan_id"]
-        else:
-            self.scan_def_ids[scan_def_id] = {"scan_id": request_block.scan_id, "point_id": 0}
-
-    def append_request_block(self, request_block: RequestBlock) -> None:
-        """append a new request block to the queue"""
-        self.request_blocks_queue.append(request_block)
-        self.request_blocks.append(request_block)
-
-    def flush_request_blocks(self) -> None:
-        """clear all request blocks from the queue"""
-        self.request_blocks = []
-        self.request_blocks_queue.clear()
-
-    def _pull_request_block(self):
-        if self.active_rb is not None:
-            return
-        if len(self.request_blocks_queue) == 0:
-            raise StopIteration
-        self.active_rb = self.request_blocks_queue.popleft()
-        self._update_point_id(self.active_rb)
-
-        self.active_rb.assign_scan_number()
-
-    def _update_point_id(self, request_block: RequestBlock):
-        if request_block.scan_def_id not in self.scan_def_ids:
-            return
-        if hasattr(request_block.scan, "point_id"):
-            if isinstance(request_block.scan.point_id, (int, float)):
-                request_block.scan.point_id = max(
-                    request_block.scan.point_id,
-                    self.scan_def_ids[request_block.scan_def_id]["point_id"],
-                )
-            else:
-                request_block.scan.point_id = self.scan_def_ids[request_block.scan_def_id][
-                    "point_id"
-                ]
-
-    def increase_scan_number(self) -> None:
-        """increase the scan number counter"""
-        rbl = self.active_rb
-        self.scan_queue.queue_manager.parent.scan_number += 1
-        if not rbl.msg.metadata.get("dataset_id_on_hold"):
-            self.scan_queue.queue_manager.parent.dataset_number += 1
-
-    def _get_metadata_for_alarm(self):
-        """get the metadata for the alarm"""
-        metadata = {}
-        if self.active_rb is None:
-            return metadata
-        if self.active_rb.scan is None:
-            return metadata
-        if self.active_rb.scan_id is not None:
-            metadata["scan_id"] = self.active_rb.scan_id
-        if self.active_rb.scan_number is not None:
-            metadata["scan_number"] = self.active_rb.scan_number
-
-        return metadata
-
-    def __iter__(self):
-        return self
-
-    def __next__(self):
-        self._pull_request_block()
-        try:
-            return next(self.active_rb.instructions)
-        except StopIteration:
-            if self.active_rb.scan_def_id in self.scan_def_ids:
-                point_id = getattr(self.active_rb.scan, "point_id", None)
-                if point_id is not None:
-                    current_point_id = self.scan_def_ids[self.active_rb.scan_def_id]["point_id"]
-                    self.scan_def_ids[self.active_rb.scan_def_id]["point_id"] = max(
-                        current_point_id, point_id
-                    )
-            self.active_rb = None
-            self._pull_request_block()
-            return next(self.active_rb.instructions)
-        except LimitError as limit_error:
-            error_info = messages.ErrorInfo(
-                error_message=limit_error.args[0],
-                compact_error_message=traceback.format_exc(limit=0),
-                exception_type=limit_error.__class__.__name__,
-                device=limit_error.device,
-            )
-            self.scan_queue.queue_manager.connector.raise_alarm(
-                severity=Alarms.MAJOR, info=error_info, metadata=self._get_metadata_for_alarm()
-            )
-            self.instruction_queue.stopped = True
-            raise ScanAbortion from limit_error
-        except DeviceInstructionError as exc:
-            logger.error(exc.message)
-            self.scan_queue.queue_manager.connector.raise_alarm(
-                severity=Alarms.MAJOR, info=exc.error_info, metadata=self._get_metadata_for_alarm()
-            )
-            raise
-        # pylint: disable=broad-except
-        except Exception as exc:
-            content = traceback.format_exc()
-            logger.error(content)
-            error_info = messages.ErrorInfo(
-                error_message=str(exc),
-                compact_error_message=traceback.format_exc(limit=0),
-                exception_type=exc.__class__.__name__,
-                device=None,
-            )
-            self.scan_queue.queue_manager.connector.raise_alarm(
-                severity=Alarms.MAJOR, info=error_info, metadata=self._get_metadata_for_alarm()
-            )
-            raise ScanAbortion from exc
-
-
-class InstructionQueueItem:
-    """The InstructionQueueItem contains and manages the request blocks for a queue item.
-    While an InstructionQueueItem can be comprised of multiple requests,
-    it will always have at max one scan_number / scan_id.
-
-    Raises:
-        StopIteration: _description_
-        StopIteration: _description_
-
-    Returns:
-        _type_: _description_
-    """
-
-    def __init__(self, parent: ScanQueue, assembler: ScanAssembler, worker) -> None:
-        self.instructions = []
-        self.parent = parent
-        self.queue = RequestBlockQueue(instruction_queue=self, assembler=assembler)
-        self.connector = self.parent.queue_manager.connector
-        self._is_scan = False
-        self.is_active = False  # set to true while a worker is processing the instructions
-        self.completed = False
-        self.exit_info: ExitInfoType | None = None
-        self.deferred_pause = True
-        self.queue_group = None
-        self.queue_group_is_closed = False
-        self.subqueue = iter([])
-        self.queue_id = str(uuid.uuid4())
-        self.scan_msgs: list[messages.ScanQueueMessage] = []
-        self.scan_assembler = assembler
-        self.worker = worker
-        self.stopped = False
-        self._status = InstructionQueueStatus.PENDING
-        self._return_to_start = None
-        self.reason: Literal["user", "alarm", "restart"] | None = None
-
-    @property
-    def scan_number(self) -> list[int]:
-        """get the scan numbers for the elements in this instruction queue"""
-        return self.queue.scan_number
-
-    @property
-    def status(self) -> InstructionQueueStatus:
-        """get the status of the instruction queue"""
-        return self._status
-
-    @status.setter
-    def status(self, val: InstructionQueueStatus) -> None:
-        """update the status of the instruction queue. By doing so, it will
-        also update its worker and publish the updated queue."""
-        logger.debug(
-            f"Setting status of instruction queue {self.queue_id} to {val.name} from thread {threading.current_thread().name}"
-        )
-        self._status = val
-        self.worker.status = val
-        if val == InstructionQueueStatus.STOPPED:
-            self.stop()
-        self.parent.queue_manager.send_queue_status()
-
-    @property
-    def active_request_block(self) -> RequestBlock:
-        """get the currently active request block"""
-        return self.queue.active_rb
-
-    @property
-    def scan_macros_complete(self) -> bool:
-        """check if the scan macro has been completed"""
-        return len(self.queue.scan_def_ids) == 0
-
-    @property
-    def scan_id(self) -> list[str]:
-        """get the scan_ids"""
-        return self.queue.scan_id
-
-    @property
-    def is_scan(self) -> list[bool]:
-        """check whether the InstructionQueue contains scan."""
-        return self.queue.is_scan
-
-    def abort(self) -> None:
-        """abort and clear all the instructions from the instruction queue"""
-        self.instructions = iter([])
-        # self.queue.request_blocks_queue.clear()
-
-    def append_scan_request(self, msg: messages.ScanQueueMessage) -> None:
-        """append a scan message to the instruction queue"""
-        self.scan_msgs.append(msg)
-        self.queue.append(msg)
-
-    def set_active(self):
-        """change the instruction queue status to RUNNING"""
-        if self.status == InstructionQueueStatus.PENDING:
-            self.status = InstructionQueueStatus.RUNNING
-
-    @property
-    def return_to_start(self) -> bool:
-        """whether or not to return to the start position after scan abortion"""
-        if self._return_to_start is not None:
-            return self._return_to_start
-        if self.active_request_block:
-            return self.active_request_block.scan.return_to_start_after_abort
-        return False
-
-    @return_to_start.setter
-    def return_to_start(self, val: bool):
-        self._return_to_start = val
-
-    def describe(self):
-        """description of the instruction queue"""
-        request_blocks = [rb.describe() for rb in self.queue.request_blocks]
-        content = messages.QueueInfoEntry(
-            queue_id=self.queue_id,
-            scan_id=self.scan_id,
-            is_scan=self.is_scan,
-            request_blocks=request_blocks,
-            scan_number=self.scan_number,
-            status=self.status.name,
-            active_request_block=(
-                self.active_request_block.describe() if self.active_request_block else None
-            ),
-            reason=self.reason or (self.exit_info[1] if self.exit_info else None),
-        )
-        return content
-
-    def append_to_queue_history(self):
-        """append a new queue item to the redis history buffer"""
-        msg = messages.ScanQueueHistoryMessage(
-            status=self.status.name, queue_id=self.queue_id, info=self.describe()
-        )
-        self.parent.queue_manager.connector.lpush(
-            MessageEndpoints.scan_queue_history(), msg, max_size=100
-        )
-
-    def __iter__(self):
-        return self
-
-    def _set_finished(self, raise_stopiteration=True):
-        self.completed = True
-        if raise_stopiteration:
-            raise StopIteration
-
-    def _get_next(
-        self, queue="instructions", raise_stopiteration=True
-    ) -> messages.DeviceInstructionMessage | None:
-        try:
-            instr = next(self.queue)
-            # instr = next(self.__getattribute__(queue))
-            if not instr:
-                return None
-            if instr.content.get("action") == "close_scan_group":
-                self.queue_group_is_closed = True
-                raise StopIteration
-            if instr.content.get("action") == "close_scan_def":
-                scan_def_id = instr.metadata.get("scan_def_id")
-                if scan_def_id in self.queue.scan_def_ids:
-                    self.queue.scan_def_ids.pop(scan_def_id)
-
-            instr.metadata["scan_id"] = self.queue.active_rb.scan_id
-            instr.metadata["queue_id"] = self.queue_id
-            self.set_active()
-            return instr
-
-        except StopIteration:
-            if not self.scan_macros_complete:
-                logger.info(
-                    "Waiting for new instructions or scan macro to be closed (scan def ids:"
-                    f" {self.queue.scan_def_ids})"
-                )
-                time.sleep(0.1)
-            elif self.queue_group is not None and not self.queue_group_is_closed:
-                self.queue.active_rb = None
-                self.parent.queue_manager.send_queue_status()
-                logger.info(
-                    "Waiting for new instructions or queue group to be closed (group id:"
-                    f" {self.queue_group})"
-                )
-                time.sleep(0.1)
-            else:
-                self._set_finished(raise_stopiteration=raise_stopiteration)
-        return None
-
-    def __next__(self):
-        if self.status in [
-            InstructionQueueStatus.RUNNING,
-            InstructionQueueStatus.DEFERRED_PAUSE,
-            InstructionQueueStatus.PENDING,
-        ]:
-            return self._get_next()
-
-        while self.status == InstructionQueueStatus.PAUSED:
-            return self._get_next(queue="subqueue", raise_stopiteration=False)
-
-        return self._get_next()
-
-    def stop(self):
-        """stop the instruction queue"""
-        blcks = self.queue.request_blocks
-        if len(blcks) > 0:
-            for blck in blcks:
-                # pylint: disable=protected-access
-                blck.scan._shutdown_event.set()
 
 
 class DirectInstructionQueueItem:
