+++ bec_server/bec_server/scan_server/scans/__init__.py
@@ -1,2 +1,1 @@
-from .legacy_scans import *
 from .scan_modifier import scan_hook
