+++ bec_server/bec_server/scan_server/scans/cont_line_scan.py
@@ -38,7 +38,7 @@ class ContLineScan(ScanBase):
     # Scan name: This is the name of the scan, e.g. "line_scan". This is used for display purposes and to identify the scan type in user interfaces.
     # Choose a descriptive name that does not conflict with existing scan names.
     # It must be a valid Python identifier, that is, it can only contain letters, numbers, and underscores, and must not start with a number.
-    scan_name = "_v4_cont_line_scan"
+    scan_name = "cont_line_scan"
     gui_config = {
         "Device": ["device", "start", "stop"],
         "Movement Parameters": ["steps", "relative", "offset", "atol"],
