+++ bec_server/bec_server/scan_server/scans/fermat_scan.py
@@ -43,7 +43,7 @@ class FermatSpiralScan(ScanBase):
     # Scan name: This is the name of the scan, e.g. "line_scan". This is used for display purposes and to identify the scan type in user interfaces.
     # Choose a descriptive name that does not conflict with existing scan names.
     # It must be a valid Python identifier, that is, it can only contain letters, numbers, and underscores, and must not start with a number.
-    scan_name = "_v4_fermat_scan"
+    scan_name = "fermat_scan"
 
     gui_config = {
         "Device 1": ["motor1", "start_motor1", "stop_motor1"],
