+++ bec_server/bec_server/scan_server/scans/multi_region_line_scan.py
@@ -36,7 +36,7 @@ class MultiRegionLineScan(ScanBase):
     # Scan name: This is the name of the scan, e.g. "line_scan". This is used for display purposes and to identify the scan type in user interfaces.
     # Choose a descriptive name that does not conflict with existing scan names.
     # It must be a valid Python identifier, that is, it can only contain letters, numbers, and underscores, and must not start with a number.
-    scan_name = "_v4_multi_region_line_scan"
+    scan_name = "multi_region_line_scan"
 
     gui_config = {
         "Movement Parameters": ["regions", "relative"],
