+++ bec_server/bec_server/scan_server/scans/round_roi_scan.py
@@ -38,7 +38,7 @@ class RoundROIScan(ScanBase):
     # Scan name: This is the name of the scan, e.g. "line_scan". This is used for display purposes and to identify the scan type in user interfaces.
     # Choose a descriptive name that does not conflict with existing scan names.
     # It must be a valid Python identifier, that is, it can only contain letters, numbers, and underscores, and must not start with a number.
-    scan_name = "_v4_round_roi_scan"
+    scan_name = "round_roi_scan"
     gui_config = {
         "Motor 1": ["motor_1", "start_motor_1", "stop_motor_1", "center_1"],
         "Motor 2": ["motor_2", "start_motor_2", "stop_motor_2", "center_2"],
