+++ bec_server/bec_server/scan_server/scans/round_scan.py
@@ -38,7 +38,7 @@ class RoundScan(ScanBase):
     # Scan name: This is the name of the scan, e.g. "line_scan". This is used for display purposes and to identify the scan type in user interfaces.
     # Choose a descriptive name that does not conflict with existing scan names.
     # It must be a valid Python identifier, that is, it can only contain letters, numbers, and underscores, and must not start with a number.
-    scan_name = "_v4_round_scan"
+    scan_name = "round_scan"
     gui_config = {
         "Motors": ["motor_1", "motor_2"],
         "Ring Parameters": [
