+++ bec_server/bec_server/scan_server/scans/time_scan.py
@@ -36,7 +36,7 @@ class TimeScan(ScanBase):
     # Scan name: This is the name of the scan, e.g. "line_scan". This is used for display purposes and to identify the scan type in user interfaces.
     # Choose a descriptive name that does not conflict with existing scan names.
     # It must be a valid Python identifier, that is, it can only contain letters, numbers, and underscores, and must not start with a number.
-    scan_name = "_v4_time_scan"
+    scan_name = "time_scan"
 
     gui_config = {"Scan Parameters": ["points", "interval", "exp_time", "settling_time"]}
